I find this description:
"Rails uses layouts as master templates. As default will you have one
master layout template called application, which you can find in
app/views/layouts/application.html.erb.

As default will this be rendered for all pages, and the content of each
page would be rendered in the yield block.

This means that application.html.erb is the right place to but generel
layout stuff, like menus and banners that should appear on all pages.

If you want to have something that varies a bit for each page (fx
different banners) can you add a special <%= yield(:banner) if
content_for?(:banner) %> in your application.html.erb file. You will
then be able to add a block in each of your pages for a banner like
this:

# app/views/some_resource/some_page.html.erb
<% content_for(:banner) do %>
  # insert page specifik banner here
<% end %>
# normal content for page
..."